refactor: move ExternalStudyRoom entity, repository
