+++ src/main/java/com/jj/swm/domain/external/study/repository/ExternalStudyRoomRepository.java
@@ -1,6 +1,6 @@
-package com.jj.swm.crawling.naver.map.repository;
+package com.jj.swm.domain.external.study.repository;
 
-import com.jj.swm.crawling.naver.map.entity.ExternalStudyRoom;
+import com.jj.swm.domain.external.study.entity.ExternalStudyRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExternalStudyRoomRepository extends JpaRepository<ExternalStudyRoom, Long> {
